==> dataset/indoor-location-navigation/Mehran Kazeminia/2-3-indoor-navigation-comparative-method.ipynb <==
<div>
    <h1 align="center"> Ensembling & Comparative Method - Part(2)</h1></h1>
    <h2 align="center">Identify the position of a smartphone in a shopping mall</h2>
    <h3 align="center">By: Somayyeh Gholami & Mehran Kazeminia</h3>
</div>

<div class="alert alert-success">  
</div>

# Description:

### - In Notebook (No. 1) we optimized the results of ten public notebooks separately. In this notebook (No. 2) we use "Ensembling" and "Comparative Method" to merge the previous results to get better scores. Finally, we publish notebooks (No. 3) called "Snap to Grid".

### - Thanks to everyone who shared their notebooks, the addresses of some of the used notebooks are as follows:

https://www.kaggle.com/robikscube/indoor-navigation-snap-to-grid-post-processing

https://www.kaggle.com/therocket290/lstm-unified-wi-fi-training-x-and-y-with-floor

https://www.kaggle.com/kokitanisaka/lstm-by-keras-with-unified-wi-fi-feats

https://www.kaggle.com/oxzplvifi/indoor-gbm-postprocessing-xy-prediction

https://www.kaggle.com/nigelhenry/simple-99-accurate-floor-model

https://www.kaggle.com/hiro5299834/wifi-features-with-lightgbm-kfold

https://www.kaggle.com/ebinan92/time-series-rnn-xy-prediction


### -  For "Generated1", we used two notebooks with scores of 5.471 and 5.995. And for the "Comparative Method" we got help from the remaining seven notebooks. If you read the code carefully, you will see that we have only allowed some of the results of these seven notebooks to affect the overall result. That is, they can only be effective if their X and Y values are between the X and Y of the two main notebooks (Rectangle for decision). This is a very important point.

### - To get "Generated2" we did exactly the same as "Generated1". But we used notebooks with scores of 5.694 and 5.995. That is another rectangle for the decision that is shared at one vertex with the previous rectangle. The rest of the steps of this notebook are to use "Ensembling".

## >>> Good Luck <<<


<div class="alert alert-success">  
</div>

# If you find this work useful, please don't forget upvoting :)

<div class="alert alert-success">  
</div>

# Import & Data Set

import numpy as np
import pandas as pd 
import matplotlib.pyplot as plt

%matplotlib inline

# _______________________________________

# Kernels Data (Public Score & File Path)

dfk = pd.DataFrame({ 
    'Kernel ID': ['A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J'],  
    'Score'    : [7.110, 6.674, 6.290, 6.189, 6.077, 6.062, 5.995, 5.694, 5.471, 5.847],  
    'File Path': ['../input/1-3-indoor-navigation-cost-minimization-floor/a9773cm99.csv', '../input/1-3-indoor-navigation-cost-minimization-floor/b9530cm99.csv', '../input/1-3-indoor-navigation-cost-minimization-floor/c8500cm99.csv', '../input/1-3-indoor-navigation-cost-minimization-floor/d8418cm99.csv', '../input/1-3-indoor-navigation-cost-minimization-floor/e8333cm99.csv', '../input/1-3-indoor-navigation-cost-minimization-floor/f8073cm99.csv', '../input/1-3-indoor-navigation-cost-minimization-floor/g7745cm99.csv', '../input/1-3-indoor-navigation-cost-minimization-floor/h7661cm99.csv', '../input/1-3-indoor-navigation-cost-minimization-floor/i7274cm99.csv', '../input/1-3-indoor-navigation-cost-minimization-floor/j7285cm99.csv']     
})    
    
dfk   

<div class="alert alert-success">  
</div>

# Ensembling & Comparative Method

## Generated(No. 1)

generated1 >>> (Public Score: ------)

ff1x = 0.60   
ff1y = 0.60  

ff2x = 0.60   
ff2y = 0.60 

ff3x = 0.90   
ff3y = 0.90 

ff4x = 0.95   
ff4y = 0.95    

main = pd.read_csv(dfk.iloc[8, 2]) 
support = pd.read_csv(dfk.iloc[6, 2])

rx , ry = [] , []
generated1  = main.copy()  

mx , my , mz = [] , [] , []
mx = main.iloc[:, 2].tolist()
my = main.iloc[:, 3].tolist()
mz = main.iloc[:, 1].tolist()

sx , sy = [] , []
sx = support.iloc[:, 2].tolist()
sy = support.iloc[:, 3].tolist()

kx = [[],[],[],[],[],[]]
ky = [[],[],[],[],[],[]]
for n in range (6):
    kernel   = pd.read_csv(dfk.iloc[n, 2])  
    kx[n] = kernel.iloc[:, 2].tolist()
    ky[n] = kernel.iloc[:, 3].tolist()

for j in range(len(main)):
    
    ex = (mx[j] * ff1x) + (sx[j] * (1.- ff1x))
    ey = (my[j] * ff1y) + (sy[j] * (1.- ff1y))
    
    if (mx[j] >= sx[j]):
        a = mx[j]
        b = sx[j]
    else: 
        a = sx[j]
        b = mx[j]
        
    if (my[j] >= sy[j]):
        c = my[j]
        d = sy[j]
    else: 
        c = sy[j]
        d = my[j]    
        
    for k in range (5, 6):            
        if ((b <= kx[k][j] <= a) and (d <= ky[k][j] <= c)):
            
            ex = (ex * ff2x) + (kx[k][j] * (1.- ff2x))
            ey = (ey * ff2y) + (ky[k][j] * (1.- ff2y))
            
            print(f'Row:{j}\t Kernel Number:{k}\t Floor:{mz[j]}')
            print(f'Main:({mx[j]},{my[j]})\nSupport:({sx[j]},{sy[j]})\nKernel:({kx[k][j]},{ky[k][j]})\nResult:({ex},{ey})\n')        
            
    for k in range (3, 5):            
        if ((b <= kx[k][j] <= a) and (d <= ky[k][j] <= c)):
            
            ex = (ex * ff3x) + (kx[k][j] * (1.- ff3x))
            ey = (ey * ff3y) + (ky[k][j] * (1.- ff3y))
            
            print(f'Row:{j}\t Kernel Number:{k}\t Floor:{mz[j]}')
            print(f'Main:({mx[j]},{my[j]})\nSupport:({sx[j]},{sy[j]})\nKernel:({kx[k][j]},{ky[k][j]})\nResult:({ex},{ey})\n')       
        
    for k in range (3):            
        if ((b <= kx[k][j] <= a) and (d <= ky[k][j] <= c)):
            
            ex = (ex * ff4x) + (kx[k][j] * (1.- ff4x))
            ey = (ey * ff4y) + (ky[k][j] * (1.- ff4y))
            
            print(f'Row:{j}\t Kernel Number:{k}\t Floor:{mz[j]}')
            print(f'Main:({mx[j]},{my[j]})\nSupport:({sx[j]},{sy[j]})\nKernel:({kx[k][j]},{ky[k][j]})\nResult:({ex},{ey})\n')
                        
    rx.append(ex)
    ry.append(ey)    

generated1.iloc[:, 2] = rx
generated1.iloc[:, 3] = ry


## Generated(No. 2)

generated2 >>> (Public Score: ------)

ff1x = 0.60   
ff1y = 0.60 

ff2x = 0.60   
ff2y = 0.60 

ff3x = 0.90   
ff3y = 0.90   

ff4x = 0.95   
ff4y = 0.95   


main = pd.read_csv(dfk.iloc[7, 2]) 
support = pd.read_csv(dfk.iloc[6, 2])

rx , ry = [] , []
generated2  = main.copy()  

mx , my , mz = [] , [] , []
mx = main.iloc[:, 2].tolist()
my = main.iloc[:, 3].tolist()
mz = main.iloc[:, 1].tolist()

sx , sy = [] , []
sx = support.iloc[:, 2].tolist()
sy = support.iloc[:, 3].tolist()

kx = [[],[],[],[],[],[]]
ky = [[],[],[],[],[],[]]
for n in range (6):
    kernel   = pd.read_csv(dfk.iloc[n, 2])  
    kx[n] = kernel.iloc[:, 2].tolist()
    ky[n] = kernel.iloc[:, 3].tolist()

for j in range(len(main)):
    
    ex = (mx[j] * ff1x) + (sx[j] * (1.- ff1x))
    ey = (my[j] * ff1y) + (sy[j] * (1.- ff1y))
    
    if (mx[j] >= sx[j]):
        a = mx[j]
        b = sx[j]
    else: 
        a = sx[j]
        b = mx[j]
        
    if (my[j] >= sy[j]):
        c = my[j]
        d = sy[j]
    else: 
        c = sy[j]
        d = my[j]    
        
    for k in range (5, 6):            
        if ((b <= kx[k][j] <= a) and (d <= ky[k][j] <= c)):
            
            ex = (ex * ff2x) + (kx[k][j] * (1.- ff2x))
            ey = (ey * ff2y) + (ky[k][j] * (1.- ff2y))
            
            print(f'Row:{j}\t Kernel Number:{k}\t Floor:{mz[j]}')
            print(f'Main:({mx[j]},{my[j]})\nSupport:({sx[j]},{sy[j]})\nKernel:({kx[k][j]},{ky[k][j]})\nResult:({ex},{ey})\n')        
        
    for k in range (3, 5):            
        if ((b <= kx[k][j] <= a) and (d <= ky[k][j] <= c)):
            
            ex = (ex * ff3x) + (kx[k][j] * (1.- ff3x))
            ey = (ey * ff3y) + (ky[k][j] * (1.- ff3y))
            
            print(f'Row:{j}\t Kernel Number:{k}\t Floor:{mz[j]}')
            print(f'Main:({mx[j]},{my[j]})\nSupport:({sx[j]},{sy[j]})\nKernel:({kx[k][j]},{ky[k][j]})\nResult:({ex},{ey})\n')           

    for k in range (3):            
        if ((b <= kx[k][j] <= a) and (d <= ky[k][j] <= c)):
            
            ex = (ex * ff4x) + (kx[k][j] * (1.- ff4x))
            ey = (ey * ff4y) + (ky[k][j] * (1.- ff4y))
            
            print(f'Row:{j}\t Kernel Number:{k}\t Floor:{mz[j]}')
            print(f'Main:({mx[j]},{my[j]})\nSupport:({sx[j]},{sy[j]})\nKernel:({kx[k][j]},{ky[k][j]})\nResult:({ex},{ey})\n')
                        
    rx.append(ex)
    ry.append(ey)    

generated2.iloc[:, 2] = rx
generated2.iloc[:, 3] = ry


## Generated(No. 3)

Use "Ensembling" on: "generated1" and "generated2"

generated3 >>> (Public Score: 5.266)

ff1x = 0.60   
ff1y = 0.60  

main = generated1
support = generated2

rx , ry = [] , []
generated3  = main.copy()  

mx , my , mz = [] , [] , []
mx = main.iloc[:, 2].tolist()
my = main.iloc[:, 3].tolist()
mz = main.iloc[:, 1].tolist()

sx , sy = [] , []
sx = support.iloc[:, 2].tolist()
sy = support.iloc[:, 3].tolist()


for j in range(len(main)):
    
    ex = (mx[j] * ff1x) + (sx[j] * (1.- ff1x))
    ey = (my[j] * ff1y) + (sy[j] * (1.- ff1y))
    
    rx.append(ex)
    ry.append(ey)    

generated3.iloc[:, 2] = rx
generated3.iloc[:, 3] = ry


## Generated(No. 4)

Use "Ensembling" on: 'I' , 'H' , 'G'

generated4 >>> (Public Score: 5.262)

ff1x = 0.44   
ff1y = 0.44  

ff2x = 0.24   
ff2y = 0.24 

ff3x = 0.32   
ff3y = 0.32 

main = pd.read_csv(dfk.iloc[8, 2])
support1 = pd.read_csv(dfk.iloc[7, 2])
support2 = pd.read_csv(dfk.iloc[6, 2])

rx , ry = [] , []
generated4 = main.copy()  

mx , my , mz = [] , [] , []
mx = main.iloc[:, 2].tolist()
my = main.iloc[:, 3].tolist()
mz = main.iloc[:, 1].tolist()

s1x , s1y = [] , []
s1x = support1.iloc[:, 2].tolist()
s1y = support1.iloc[:, 3].tolist()

s2x , s2y = [] , []
s2x = support2.iloc[:, 2].tolist()
s2y = support2.iloc[:, 3].tolist()


for j in range(len(main)):
    
    ex = (mx[j] * ff1x) + (s1x[j] * ff2x) + (s2x[j] * ff3x)
    ey = (my[j] * ff1y) + (s1y[j] * ff2y) + (s2y[j] * ff3y)
    
    rx.append(ex)
    ry.append(ey)    

generated4.iloc[:, 2] = rx
generated4.iloc[:, 3] = ry


## Generated(No. 5)

Use "Ensembling" on: "generated3" and "generated4"

generated5 >>> (Public Score: 5.255)

ff1x = 0.50   
ff1y = 0.50  

main = generated3
support = generated4

rx , ry = [] , []
generated5  = main.copy()  

mx , my , mz = [] , [] , []
mx = main.iloc[:, 2].tolist()
my = main.iloc[:, 3].tolist()
mz = main.iloc[:, 1].tolist()

sx , sy = [] , []
sx = support.iloc[:, 2].tolist()
sy = support.iloc[:, 3].tolist()


for j in range(len(main)):
    
    ex = (mx[j] * ff1x) + (sx[j] * (1.- ff1x))
    ey = (my[j] * ff1y) + (sy[j] * (1.- ff1y))
    
    rx.append(ex)
    ry.append(ey)    

generated5.iloc[:, 2] = rx
generated5.iloc[:, 3] = ry


## Generated(No. 6)

Use "Ensembling" on: 'I' , 'H' , 'G'

generated6 >>> (Public Score: 5.265)

ff1x = 0.40   
ff1y = 0.40  

ff2x = 0.25   
ff2y = 0.25 

ff3x = 0.35   
ff3y = 0.35 

main = pd.read_csv(dfk.iloc[8, 2])
support1 = pd.read_csv(dfk.iloc[7, 2])
support2 = pd.read_csv(dfk.iloc[6, 2])

rx , ry = [] , []
generated6 = main.copy()  

mx , my , mz = [] , [] , []
mx = main.iloc[:, 2].tolist()
my = main.iloc[:, 3].tolist()
mz = main.iloc[:, 1].tolist()

s1x , s1y = [] , []
s1x = support1.iloc[:, 2].tolist()
s1y = support1.iloc[:, 3].tolist()

s2x , s2y = [] , []
s2x = support2.iloc[:, 2].tolist()
s2y = support2.iloc[:, 3].tolist()


for j in range(len(main)):
    
    ex = (mx[j] * ff1x) + (s1x[j] * ff2x) + (s2x[j] * ff3x)
    ey = (my[j] * ff1y) + (s1y[j] * ff2y) + (s2y[j] * ff3y)
    
    rx.append(ex)
    ry.append(ey)    

generated6.iloc[:, 2] = rx
generated6.iloc[:, 3] = ry


<div class="alert alert-success">  
</div>

# Submission

sub = generated5
sub.to_csv("submission.csv", index=False)

generated1.to_csv("generated1.csv", index=False)
generated2.to_csv("generated2.csv", index=False)
generated3.to_csv("generated3.csv", index=False)
generated4.to_csv("generated4.csv", index=False)
generated5.to_csv("generated5.csv", index=False)
generated6.to_csv("generated6.csv", index=False)

!ls

<div class="alert alert-success">  
</div>
